+++ CHANGELOG.md
@@ -14,12 +14,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 - The `previewer` script no longer skips non-regular files (#2327).
 - Line numbers now take up less space when both `number` and `relativenumber` are enabled (#2331).
+- Changes have been made to ruler configuration (#2338):
+  - The ruler file is no longer experimental.
+  - The ruler file will be used by default unless `rulerfmt` (now a blank string by default) is specified.
+  - The ruler file is no longer read from fixed locations like `~/.config/lf/ruler`, and instead the `rulerfile` option has been repurposed to specify the path of the ruler file.
 
 ### Added
 
 - A new server command `list` is added to print the IDs of all currently connected clients (#2314).
 - The `previewer` and `cleaner` scripts now have their `stderr` output logged (#2316).
 - A new option `mergeindicators` is added to reduce the gap before filenames, by merging tag and selection indicators into a single column (#2330).
+- The `ruler` file now supports `.Stat.Extension`, `.Stat.AccessTime`, `.Stat.BirthTime`, `.Stat.ChangeTime` and `.Stat.CustomInfo` (#2329), as well as `.Stat.DirSize` and `.Stat.DirCount` (#2343).
 
 ### Fixed
 
@@ -27,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Newline characters are now ignored when drawing the ruler with the `ruler` file configured (#2319).
 - A potential crash when using the `scroll-up`/`scroll-down` commands is now fixed (#2320).
 - Case-insensitive command-line completions no longer cause user input to be displayed in lowercase (#2336).
+- Calculation of window widths for the `ratios` option is now more accurate (#2347).
 
 ## [r40](https://github.com/gokcehan/lf/releases/tag/r40)